+++ js/common/message.js
@@ -1,0 +1,6 @@
+export const messages = {
+  emptyEmail: "이메일을 입력해 주세요.",
+  invalidEmail: "올바른 이메일 주소가 아닙니다.",
+  emptyPassword: "비밀번호를 입력해 주세요.",
+  incorrectCredentials: "이메일 또는 비밀번호를 확인해 주세요.",
+};
