+++ js/common/ui.js
@@ -1,0 +1,4 @@
+export function togglePassword(pwInput, eyeButton) {
+  pwInput.type = pwInput.type === "password" ? "text" : "password";
+  eyeButton.classList.toggle("eye-button-off");
+}
