+++ pages/signin.html
@@ -14,9 +14,11 @@
   <body>
     <header>
       <a class="logo-link" href="../index.html"
-        ><img src="../images/signin/logo.svg" alt="Linkbrary" /
-        class="header-logo"></a
-      >
+        ><img
+          src="../images/signin/logo.svg"
+          alt="Linkbrary"
+          class="header-logo"
+      /></a>
       <p class="header-message">
         회원이 아니신가요?
         <a class="header-link" href="./signup.html">회원 가입하기</a>
@@ -64,6 +66,6 @@
         </div>
       </div>
     </div>
-    <script src="../js/signin.js"></script>
+    <script type="module" src="/js/siginin.js"></script>
   </body>
 </html>
